+++ src/cpp/session/modules/SessionLSP.hpp
@@ -276,6 +276,7 @@ struct Events : boost::noncopyable
    RSTUDIO_BOOST_SIGNAL<void(DidOpenTextDocumentParams)>   didOpen;
    RSTUDIO_BOOST_SIGNAL<void(DidChangeTextDocumentParams)> didChange;
    RSTUDIO_BOOST_SIGNAL<void(DidCloseTextDocumentParams)>  didClose;
+   RSTUDIO_BOOST_SIGNAL<void(DidFocusTextDocumentParams)>  didFocus;
 };
 
 inline Events& events()
@@ -286,6 +287,13 @@ inline Events& events()
 
 core::Error sourceDocumentFromUri(
    const std::string& uri,
+   boost::shared_ptr<source_database::SourceDocument> pDoc);
+
+// As above, but allows skipping the read of the document's contents
+// sidecar file when only document metadata is needed.
+core::Error sourceDocumentFromUri(
+   const std::string& uri,
+   bool includeContents,
    boost::shared_ptr<source_database::SourceDocument> pDoc);
 
 std::string uriFromDocument(boost::shared_ptr<source_database::SourceDocument> pDoc);